Add information about the SOK battery I bought
I will use this battery to replace the SLA I currently have. I can reuse the solar charge controller I am using with it for now, but in the future, I will need to get a better one for sure. 

I have added an image which I will upload next.